Scanning multiple sites in one pipeline.
The Concourse S3 resource does not support uploading multiple files. This commit switches the S3 uploading to be handled in process, instead of by Concourse itself.

The list of targets comes from a json file in this repository (targets.json), and the result files are written to results/{name of app}.json.