Switcing to use the latest slack notification resource.
Defining the `resource_type` for the Slack notification caused the pipeline to pull the outdated version from docker hub (https://hub.docker.com/r/nopik/slack-notification-resource/) instead of the built-in version we're using (https://github.com/18F/slack-notification-resource-boshrelease). There is an open issue for this here: https://github.com/cloudfoundry-community/slack-notification-resource/issues/10

Additionally, I moved the slack notification to the end of list of tasks. It was erroring after it notified the user of success, but before the files were actually uploaded, which is probably the worst way for this to error out. :)