Downcasing target and project names before using them.
We ran into an issue with `C2`, which was specified in the targets.json file as `c2`, in the team-api as `C2`, and as a team-api url as `c2`. Since the team-api urls are downcased, it makes the most sense to standardize on that format here.

Just downcasing them should prevent tasks further along in the pipeline (or compliance-viewer) from needing to think about this.

Addresses #33