Removing reliance on zap-cli creating new sessions.
This is related to https://github.com/18F/concourse-compliance-testing/issues/48 . Creating new sessions seems unreliable if the session returns a large amount of data. This goes back to starting/shutting down ZAP for each site. Starting/shutting down ZAP is slower, but that is less of an issue now that we are one project per job.